Renamed vector_operation.py and matrix_operation.py to plural forms
